fix:rename frontamtter to frontmatter
