+++ scripts/create-i18n-tree.mjs
@@ -1,7 +1,7 @@
 import { existsSync } from "fs";
 import path from "path";
 import { COLLECTIONS_ROOT } from "./collections/index.mjs";
-import { getFrontMatterData } from "./collections/utils/get-frontamtter.mjs";
+import { getFrontMatterData } from "./collections/utils/get-frontmatter.mjs";
 
 /**
  * @typedef {Object} Leaf
